Move connection_status tests into their own module.
